requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
fake-useragent==0.1.11
idna==2.10
lxml==4.6.1
Pillow==8.0.1
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.11
